DEV-7 integrated component with the request and moved it to containers
